+++ packages/cli/plugin-bff/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @modern-js/plugin-bff
 
+## 1.21.1
+
+### Patch Changes
+
+- @modern-js/bff-core@1.21.1
+- @modern-js/create-request@1.21.1
+- @modern-js/server-utils@1.21.1
+- @modern-js/babel-compiler@1.21.1
+- @modern-js/utils@1.21.1
+
 ## 1.21.0
 
 ### Patch Changes