+++ packages/cli/plugin-jarvis/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/plugin-jarvis
 
+## 1.21.1
+
+### Patch Changes
+
+- @modern-js/eslint-config@1.21.1
+- @modern-js-app/eslint-config@1.21.1
+- @modern-js/tsconfig@1.21.1
+- @modern-js/utils@1.21.1
+
 ## 1.21.0
 
 ### Patch Changes