+++ packages/server/core/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @modern-js/server-plugin
 
+## 1.21.1
+
+### Patch Changes
+
+- @modern-js/plugin@1.21.1
+- @modern-js/utils@1.21.1
+
 ## 1.21.0
 
 ### Patch Changes